+++ src/lib.rs
@@ -110,11 +110,15 @@ mod span;
 
 #[cfg(not(feature = "chrono"))]
 pub use crate::date::SimpleDate;
-pub use crate::date::{Date, DateCompound};
-pub use crate::description::{Component, Components, Description};
-pub use crate::priority::Priority;
-pub use crate::state::State;
-pub use crate::task::{Task, TaskBuilder};
+pub use crate::date::{
+	Date, DateCompound, ParseDateCompoundError, ParseDateError,
+};
+pub use crate::description::{
+	Component, Components, Description, ParseDescriptionError,
+};
+pub use crate::priority::{ParsePriorityError, Priority};
+pub use crate::state::{ParseStateError, State};
+pub use crate::task::{ParseTaskError, Task, TaskBuilder};
 
 pub mod prelude {
 	//! The prelude exports all components needed for regular use.
@@ -127,11 +131,15 @@ pub mod prelude {
 
 	#[cfg(not(feature = "chrono"))]
 	pub use crate::date::SimpleDate;
-	pub use crate::date::{Date, DateCompound};
-	pub use crate::description::{Component, Components, Description};
-	pub use crate::priority::Priority;
-	pub use crate::state::State;
-	pub use crate::task::{Task, TaskBuilder};
+	pub use crate::date::{
+		Date, DateCompound, ParseDateCompoundError, ParseDateError,
+	};
+	pub use crate::description::{
+		Component, Components, Description, ParseDescriptionError,
+	};
+	pub use crate::priority::{ParsePriorityError, Priority};
+	pub use crate::state::{ParseStateError, State};
+	pub use crate::task::{ParseTaskError, Task, TaskBuilder};
 }
 
 #[cfg(test)]